Add files via upload
When you are using multiple servos that are under constant load, without sufficient power it will cause the microcontroller to reset. If for some reason you cannot immediately fix the electrical need, this detach example shows how to use one servo at a time by detaching and reattaching servos. This will send no signal to the detached servo and treat it as if it were unplugged.